Update forgetting_mlp.py
I tried to ran a few extra tests using different losses, optimizers, and dropout settings to see if any of them reduced forgetting. This was not part of the main task but thought it might help me understand how training choices influence memory retention. But this took way too long to run so decided not to run.